Deepfake-Detection/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Flask==3.1.3
flask-cors==6.0.2
supabase==2.28.0
firebase-admin==7.2.0
python-dotenv==1.2.2
torch==2.9.1+cpu
torchvision==0.24.1+cpu
opencv-python-headless
transformers
google-generativeai
